ci(hooks): add file-size pre-push gate
